fix:전체 레이아웃 적용
