--- src/app/(with-routes)/(with-protected)/profile/client/page.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function ClientProfilePage() {
-  return <div>ClientProfilePage</div>;
-}
-
-export default ClientProfilePage;
