--- src/app/(with-routes)/(with-public)/example/page.tsx
@@ -1,15 +1,0 @@
-//컴포넌트 테스트용 페이지입니다.
-//배포 단계에선 삭제할 예정입니다.
-
-import React from "react";
-import MoverProfilePage from "../../(with-protected)/profile/mover/page";
-
-function EamplePage() {
-  return (
-    <>
-      <MoverProfilePage />
-    </>
-  );
-}
-
-export default EamplePage;
