+++ src/lib/validations/index.ts
@@ -1,1 +1,2 @@
-export * from './common/email.validations'
+export * from "./common/email.validations";
+export * from "./profile.validations";
